tests/wpt/web-platform-tests/trusted-types/HTMLElement-generic.tentative.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/helper.sub.js"></script>
</head>
<body>
<script>
  var testnb = 0;
  // TrustedURL Assignments
  const URLTestCases = [
    [ 'a', 'href' ],
    [ 'area', 'href' ],
    [ 'base', 'href' ],
    [ 'frame', 'src' ],
    [ 'iframe', 'src' ],
    [ 'img', 'src' ],
    [ 'input', 'src' ],
    [ 'link', 'href' ],
    [ 'video', 'src' ],
    [ 'object', 'data' ],
    [ 'object', 'codeBase' ],
    [ 'source', 'src' ],
    [ 'track', 'src' ]
  ];

  URLTestCases.forEach(c => {
    test(t => {
      assert_element_accepts_trusted_url(window, ++testnb, t, c[0], c[1], RESULTS.URL);
    }, c[0] + "." + c[1] + " assigned via policy (successful URL transformation)");
  });

  // TrustedScriptURL Assignments
  const scriptURLTestCases = [
    [ 'embed', 'src' ],
    [ 'script', 'src' ]
  ];

  testnb = 0;
  scriptURLTestCases.forEach(c => {
    test(t => {
      assert_element_accepts_trusted_script_url(window, ++testnb, t, c[0], c[1], RESULTS.SCRIPTURL);
    }, c[0] + "." + c[1] + " assigned via policy (successful ScriptURL transformation)");
  });

  // TrustedHTML Assignments
  const HTMLTestCases = [
    [ 'div', 'innerHTML' ],
    [ 'iframe', 'srcdoc' ]
  ];

  testnb = 0;
  HTMLTestCases.forEach(c => {
    test(t => {
      assert_element_accepts_trusted_html(window, ++testnb, t, c[0], c[1], RESULTS.HTML);
    }, c[0] + "." + c[1] + " assigned via policy (successful HTML transformation)");
  });

  // TrustedScript Assignments
  const scriptTestCases = [
    [ 'script', 'text' ],
    [ 'script', 'innerText' ],
    [ 'script', 'textContent' ]
  ];

  testnb = 0;
  scriptTestCases.forEach(c => {
    test(t => {
      assert_element_accepts_trusted_script(window, c, t, c[0], c[1], RESULTS.SCRIPT);
    }, c[0] + "." + c[1] + " assigned via policy (successful Script transformation)");
  });
</script>
